Restructure spawning module
